usb: Rework the USB core internals.
Rework the way how the interfaces are handled by the USB core stack.
Until now USB stack was aware of only the first interface associated
with each class. This led to a problem when not first interface was
addressed by the request.

This patch allow to filter out the request at core level.

The usb_cfg_data struct is legacy of initial implementation
of USB device stack. Initially the USB device stack was
implemented to support single function devices with
no composite support. This structure was intended to be used to
configure the device in run time by calling usb_set_config().

Because of how things grow the structure was used to represent
the functions instead of devices.

This patch removes device configuration descriptor pointer.
Adds pointer to table of USB interface containers associated with
he function it represents. Each container consist of pointer to
interface descriptor, its alternate setting (if exists) and currently
selected alternate setting. This help USB core stack to properly address
SET/GET interface requests.

This patch should be followed by renaming the structure to something
closer to 'function' specific data as described in USB 2.0 spec.

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>